requirements.txt
pytest>=7.0
python-dotenv>=1.0
sentence-transformers>=3.0
openai>=1.0
chromadb>=0.4.0
ragas
unstructured
python-docx
pymupdf
pymupdf4llm
pandas
streamlit
